+++ app/webpage/sources/components/screen/start/StartScreen.js
@@ -21,7 +21,8 @@ export default class StartScreen extends React.Component {
       loading: true,
       aiDefList: [],
       battleSet: battleSet,
-      teamMode: teamMode
+      teamMode: teamMode,
+      ubdErrorMessage: null
     };
     this.difficultyMap = {};
     this.difficultyMap.dummy = 1;
@@ -101,7 +102,7 @@ export default class StartScreen extends React.Component {
   startBattle() {
     let listComplete = this.state.aiDefList.length >= 2;
     if(!listComplete) return;
-    this.props.onStart(this.state.aiDefList, {teamMode: this.state.teamMode});
+    this.props.onStart(this.state.aiDefList, {teamMode: this.state.teamMode, rngSeed: Math.random()});
   }
 
   onSettingsChange(tankName, v) {
@@ -204,6 +205,39 @@ export default class StartScreen extends React.Component {
     });
   }
 
+  onUbdLoad(event) {
+    let file = event.target.files[0];
+    let reader = new FileReader();
+    let self = this;
+    reader.onload = function (e) {
+      let content = e.target.result;
+      let descriptor = JsBattle.createUBD();
+      try {
+        content = content.split("base64,")[1];
+        content = atob(content);
+        descriptor.decode(content);
+      } catch (err) {
+        console.log(err);
+        self.setState({
+          ubdErrorMessage: "Error! Cannot parse *.UBD file!",
+          rngSeed: Math.random()
+        });
+        return;
+      }
+      self.setState({
+        ubdErrorMessage: null
+      });
+      self.props.onStart(
+        descriptor.getAiList(),
+        {
+          teamMode: descriptor.getTeamMode(),
+          rngSeed: descriptor.getRngSeed()
+        }
+      );
+    };
+    reader.readAsDataURL(event.target.files[0]);
+  }
+
   renderSettings() {
     return <div>
       <button type="button" className="btn btn-success btn-lg pull-right create-tank" onClick={() => this.createTank()} style={{margin: "15px"}}>
@@ -235,6 +269,13 @@ export default class StartScreen extends React.Component {
 
   render() {
     let listComplete = this.state.aiDefList.length >= 2;
+    let ubdError = <div className="alert alert-danger" role="alert" style={{marginTop: "5px"}}>
+      <i class="fa fa-exclamation-circle" aria-hidden="true"></i> {this.state.ubdErrorMessage}
+    </div>;
+    if(!this.state.ubdErrorMessage) {
+      ubdError = null;
+    }
+
     let content = <Row>
       <Col lg={4} md={5}>
         <div className="card">
@@ -251,6 +292,24 @@ export default class StartScreen extends React.Component {
                Enable <strong>Team Mode</strong> for tanks of the same type
              </label>
            </div>
+          </div>
+        </div>
+        <div className="card text-white bg-dark" style={{marginTop: '10px'}}>
+          <div className="card-body">
+            <h5 className="card-title">Ultimate Battle Descriptor</h5>
+            <p className="card-text">
+              UBD files contain all the information required to replay the battle. If you have an *.UBD file,
+              load it here to re-watch the competition.
+              <input
+                onChange={(e) => this.onUbdLoad(e)}
+                className="btn btn-light"
+                style={{marginTop: "20px", width: "100%"}}
+                type="file" id="ubdUpload"
+                name="ubdFile"
+                accept=".ubd"
+              />
+              {ubdError}
+            </p>
           </div>
         </div>
       </Col>