+++ app/webpage/sources/components/screen/winner/WinnerScreen.js
@@ -13,6 +13,10 @@ export default class WinnerScreen extends React.Component {
     return result.teamMode ? "Team " + result.teamWinner.name : result.tankWinner.fullName;
   }
 
+  getUBD() {
+    return this.props.result.ubd;
+  }
+
   getWinnerSkin() {
     return this.props.result.tankWinner.skin;
   }
@@ -51,7 +55,7 @@ export default class WinnerScreen extends React.Component {
 
   render() {
     let previewUrl = "img/tank_skin_" + this.getWinnerSkin() + ".png";
-
+    let ubdUri = 'data:text/plain;base64,' + btoa(this.getUBD());
 
     let restartButton = <button className="btn btn-secondary btn-lg" onClick={() => this.props.onRestart()}>
       <i className="fa fa-refresh" aria-hidden="true"></i> Next Battle
@@ -78,6 +82,13 @@ export default class WinnerScreen extends React.Component {
               &nbsp;
               {editButton}
             </div>
+          </div>
+        </div>
+        <div className="card text-white bg-dark" style={{marginTop: '10px'}}>
+          <div className="card-body">
+            <h5 className="card-title">Ultimate Battle Descriptor</h5>
+            <p className="card-text">UBD file contains all the information required to replay the battle later.</p>
+            <a href={ubdUri} download="battle.ubd" className="btn btn-light" style={{width: "100%"}}>Download <strong>battle.ubd</strong> file</a>
           </div>
         </div>
       </Col>